Title: 배열의 특정 위치 찾기, Time: 7ms, Memory: 0MB, Status: Passed - CodeTree

+++ 240724/배열의 특정 위치 찾기/find-specific-location-fo-array.c
@@ -14,7 +14,7 @@ int main() {
 
         if((i+1)%3==0)
         sum1+=arr[i];
-        avg=sum1/3;
+        avg=(double)sum1/3;
     }
 
     printf("%d %.1lf", sum, avg);
